+++ src/core/lv_obj_class.c
@@ -42,7 +42,7 @@ static uint32_t get_instance_size(const lv_obj_class_t * class_p);
 
 lv_obj_t * lv_obj_class_create_obj(const lv_obj_class_t * class_p, lv_obj_t * parent)
 {
-    LV_TRACE_OBJ_CREATE("Creating object with %p class on %p parent", class_p, parent);
+    LV_TRACE_OBJ_CREATE("Creating object with %p class on %p parent", (void *)class_p, (void *)parent);
     uint32_t s = get_instance_size(class_p);
     lv_obj_t * obj = lv_mem_alloc(s);
     if(obj == NULL) return NULL;
@@ -123,6 +123,7 @@ void lv_obj_class_init_obj(lv_obj_t * obj)
         /*Call the ancestor's event handler to the parent to notify it about the new child.
          *Also triggers layout update*/
         lv_event_send(parent, LV_EVENT_CHILD_CHANGED, obj);
+        lv_event_send(parent, LV_EVENT_CHILD_CREATED, obj);
 
         /*Invalidate the area if not screen created*/
         lv_obj_invalidate(obj);